update - move get_schema to tests
